+++ coronavirus_map/cli.py
@@ -7,8 +7,6 @@ import click
 import coronavirus_map.application.map_populator as map_populator
 import coronavirus_map.domain.classifier as classifier
 import coronavirus_map.domain.mapper as mapper
-import training_scripts.application.data_generation_service as data_generation_service
-import training_scripts.domain.sampler as sampler
 
 
 @click.command()
